server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
starlette==0.37.2
onnxruntime==1.18.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.3.0
qrcode==7.4.2
